created App.js and imported to index.js
